+++ include/CheckRegister.hpp
@@ -1,0 +1,10 @@
+#ifndef CHECKREGISTER_HPP
+#define CHECKREGISTER_HPP
+
+#include "Client.hpp"
+#include "Message.hpp"
+
+bool check_authenticated(Client &client);
+bool check_seted_nick_user(Client &client);
+
+#endif // CHECKREGISTER_HPP
